[TEST][#64] Move test from Client to Compiler [#950]
