+++ Compiler/TranslatorTests/TestProjects/16/Bridge/reference/test.bridgeIssues.n064.js
@@ -1,0 +1,22 @@
+﻿/* global Bridge */
+
+(function (globals) {
+    "use strict";
+
+    Bridge.define('Test.BridgeIssues.N064.Class64', {
+        constructor: function () {
+        },
+        constructor$1: function (related) {
+        },
+        test: function () {
+            var aux = new Test.BridgeIssues.N064.Class64.Aux1();
+            new Test.BridgeIssues.N064.Class64("constructor$1", aux);
+        }
+    });
+    
+    Bridge.define('Test.BridgeIssues.N064.Class64.Aux1');
+    
+    
+    
+    Bridge.init();
+})(this);
