+++ Compiler/TranslatorTests/TestProjects/16/test.csproj
@@ -34,6 +34,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="BridgeIssues\N059.cs" />
+    <Compile Include="BridgeIssues\N064.cs" />
     <Compile Include="BridgeIssues\N062.cs" />
     <Compile Include="BridgeIssues\N447.cs" />
     <Compile Include="BridgeIssues\N527.cs" />
