+++ Tests/JavaScript/Batch1/Tests/NotCompiling.cs
@@ -24,28 +24,6 @@ namespace Bridge.ClientTest.BridgeIssues
     //     }
     // }
 
-    // [#64]
-    internal class Class64
-    {
-        public class Aux1
-        {
-        }
-
-        public Class64()
-        {
-        }
-
-        public Class64(Aux1 related)
-        {
-        }
-
-        public void Test()
-        {
-            var aux = new Class64.Aux1();
-            new Class64(aux);
-        }
-    }
-
     // [#65]
     internal class Class65_1
     {
